Refactored the documentation, mostly the tutorials, and deprecated some classes and methods
