+++ docs/api/benchmark.md
@@ -1,16 +1,10 @@
-# Base class
+::: polaris.benchmark.BenchmarkV2Specification
+    options:
+        filters: ["!^_", "!md5sum", "!get_cache_path"]
+
 
 ::: polaris.benchmark.BenchmarkSpecification
     options:
         filters: ["!^_", "!md5sum", "!get_cache_path"]
-
----
-## Subclasses
-
-::: polaris.benchmark.SingleTaskBenchmarkSpecification
-
----
-
-::: polaris.benchmark.MultiTaskBenchmarkSpecification
 
 ---
