+++ docs/quickstart.md
@@ -1,39 +1,48 @@
 # Quickstart
+Welcome to the Polaris Quickstart guide! This page will introduce you to core concepts and you'll submit a first result to a benchmark on the [Polaris Hub](https://www.polarishub.io).
 
 ## Installation
+!!! warning "`polaris-lib` vs `polaris`"
+    Be aware that the package name differs between _pip_ and _conda_.
 
-First things first, let's install Polaris!
-
-We highly recommend using a [Conda Python distribution](https://github.com/conda-forge/miniforge), such as `mamba`:
+Polaris can be installed via _pip_:
 
 ```bash
-mamba install -c conda-forge polaris
+pip install polaris-lib
 ```
 
-??? info "Other installation options"
-    You can replace `mamba` by `conda`. The package is also pip installable if you need it: `pip install polaris-lib`.
+or _conda_: 
+```bash
+conda install -c conda-forge polaris
+```
 
-## Authenticating to the Polaris Hub
-To interact with the [Polaris Hub](https://polarishub.io/) from the client, you must first login. You can do this
-via the following command in your terminal:
+## Core concepts
+Polaris explicitly distinguished **datasets** and **benchmarks**. 
+
+- A _dataset_ is simply a tabular collection of data, storing datapoints in a row-wise manner. 
+- A _benchmark_ defines the ML task and evaluation logic (e.g. split and metrics) for a dataset.
+
+One dataset can therefore be associated with multiple benchmarks. 
+
+## Login
+To interact with the [Polaris Hub](https://polarishub.io/) from the client, you must first authenticate yourself. If you don't have an account yet, you can create one [here](https://polarishub.io/sign-up).
+
+You can do this via the following command in your terminal:
 
 ```bash
 polaris login
 ```
 
-This will redirect you to a login page on the Polaris Hub where you can either sign in or sign up. Once either
-of these options have been completed, you will see an authorization code on your screen. Copy this and paste it 
-back into your terminal when prompted by the client.
+or in Python: 
+```py
+from polaris.hub.client import PolarisHubClient
 
-That's it! You're now all set to interact with datasets and benchmarks across Polaris.
+with PolarisHubClient() as client:
+    client.login()
+```
 
-## Benchmarking API
-
-At its core, Polaris is a benchmarking library. It provides a simple API to run benchmarks. While it can be used
-independently, it is built to easily integrate with the Polaris Hub. The hub hosts
-a variety of high-quality datasets, benchmarks and associated results.
-
-If all you care about is to partake in a benchmark that is hosted on the hub, it is as simple as:
+## Benchmark API
+To get started, we will submit a result to the [`polaris/hello-world-benchmark`](https://polarishub.io/benchmarks/polaris/hello-world-benchmark).
 
 ```python
 import polaris as po
@@ -57,17 +66,18 @@ predictions = [0.0 for x in test]
 results = benchmark.evaluate(predictions)
 
 # Submit your results
-results.upload_to_hub(owner="dummy-user")
+results.upload_to_hub(owner="dummy-user", access="public")
 ```
 
-That's all there is to it to partake in a benchmark. No complicated, custom data-loaders or evaluation protocol. With just a few lines of code, you can feel confident that you are properly evaluating your model and focus on what you do best: Solving the hard problems in our domain!
+Through immutable datasets and standardized benchmarks, Polaris aims to serve as a source of truth for machine learning in drug discovery. The limited flexibility might differ from your typical experience, but this is by design to improve reproducibility. Learn more [here](https://polarishub.io/blog/reproducible-machine-learning-in-drug-discovery-how-polaris-serves-as-a-single-source-of-truth).
 
-Similarly, you can easily access a dataset.
+## Dataset API
+Loading a benchmark will automatically load the underlying dataset. We can also directly access the [`polaris/hello-world`](https://polarishub.io/datasets/polaris/hello-world) dataset.
 
 ```python
 import polaris as po
 
-# Load the dataset from the hub
+# Load the dataset from the Hub
 dataset = po.load_dataset("polaris/hello-world")
 
 # Get information on the dataset size
@@ -82,21 +92,14 @@ dataset.get_data(
 # Or, similarly:
 dataset[dataset.rows[0], dataset.columns[0]]
 
-# Get an entire row
+# Get an entire data point
 dataset[0]
 ```
 
-## Core concepts
-
-At the core of our API are 4 core concepts, each associated with a class:
-
-1. [`Dataset`][polaris.dataset.Dataset]: The dataset class is carefully designed data-structure, stress-tested on terra-bytes of data, to ensure whatever dataset you can think of, you can easily create, store and use it.
-2. [`BenchmarkSpecification`][polaris.benchmark.BenchmarkSpecification]: The benchmark specification class wraps a `Dataset` with additional meta-data to produce a the benchmark. Specifically, it specifies how to evaluate a model's performance on the underlying dataset (e.g. the train-test split and metrics). It provides a simple API to run said evaluation protocol.
-3. [`Subset`][polaris.dataset.Subset]: The subset class should be used as a starting-point for any framework-specific (e.g. PyTorch or Tensorflow) data loaders. To facilitate this, it abstracts away the non-trivial logic of accessing the data and provides several style of access to built upon.
-4. [`BenchmarkResults`][polaris.evaluate.BenchmarkResults]: The benchmark results class stores the results of a benchmark, along with additional meta-data. This object can be easily uploaded to the Polaris Hub and shared with the broader community.
+Drug discovery research involves a maze of file formats (e.g. PDB for 3D structures, SDF for small molecules, and so on). Each format requires specialized knowledge to parse and interpret properly. At Polaris, we wanted to remove that barrier. We use a universal data format based on [Zarr](https://zarr.dev/). Learn more [here](https://polarishub.io/blog/dataset-v2-built-to-scale).
 
 ## Where to next?
 
-Now that you've seen how easy it is to use Polaris, let's dive into the details through a set of tutorials!
+Now that you've seen how easy it is to use Polaris, let's dive into the details through [a set of tutorials](./tutorials/submit_to_benchmark.ipynb)!
 
 ---
