+++ polaris/benchmark/__init__.py
@@ -1,9 +1,11 @@
 from polaris.benchmark._base import BenchmarkSpecification, BenchmarkV1Specification
+from polaris.benchmark._benchmark_v2 import BenchmarkV2Specification
 from polaris.benchmark._definitions import MultiTaskBenchmarkSpecification, SingleTaskBenchmarkSpecification
 
 __all__ = [
     "BenchmarkSpecification",
     "BenchmarkV1Specification",
+    "BenchmarkV2Specification",
     "SingleTaskBenchmarkSpecification",
     "MultiTaskBenchmarkSpecification",
 ]
