+++ polaris/benchmark/_base.py
@@ -4,6 +4,7 @@ from hashlib import md5
 from itertools import chain
 from pathlib import Path
 from typing import Any, Callable, ClassVar, Literal
+from warnings import deprecated
 
 import fsspec
 import numpy as np
@@ -75,18 +76,13 @@ class BaseSplitSpecificationMixin(BaseModel):
 class BenchmarkSpecification(
     PredictiveTaskSpecificationMixin, BaseArtifactModel, BaseSplitSpecificationMixin, abc.ABC
 ):
-    """This class wraps a [`Dataset`][polaris.dataset.Dataset] with additional data
-     to specify the evaluation logic.
+    """This class wraps a dataset with additional data to specify the evaluation logic.
 
     Specifically, it specifies:
 
-    1. Which dataset to use (see [`Dataset`][polaris.dataset.Dataset]);
+    1. Which dataset to use;
     2. A task definition (we currently only support predictive tasks);
     3. A predefined, static train-test split to use during evaluation.
-
-    info: Subclasses
-        Polaris includes various subclasses of the `BenchmarkSpecification` that provide a more precise data-model or
-         additional logic, e.g. [`SingleTaskBenchmarkSpecification`][polaris.benchmark.SingleTaskBenchmarkSpecification].
 
     Examples:
         Basic API usage:
@@ -236,7 +232,6 @@ class BenchmarkSpecification(
         Warning: Multiple files
             Perhaps unintuitive, this method creates multiple files in the destination directory as it also saves
             the dataset it is based on to the specified destination.
-            See the docstring of [`Dataset.to_json`][polaris.dataset.Dataset.to_json] for more information.
 
         Args:
             destination: The _directory_ to save the associated data to.
@@ -273,6 +268,7 @@ class BenchmarkSpecification(
         return self.__repr__()
 
 
+@deprecated("Use BenchmarkV2Specification instead")
 class BenchmarkV1Specification(SplitSpecificationV1Mixin, ChecksumMixin, BenchmarkSpecification):
     _version: ClassVar[Literal[1]] = 1
