+++ polaris/benchmark/_benchmark_v2.py
@@ -4,8 +4,8 @@ from pydantic import Field, field_validator, model_validator
 from typing_extensions import Self
 
 from polaris.benchmark import BenchmarkSpecification
+from polaris.benchmark._split_v2 import SplitSpecificationV2Mixin
 from polaris.dataset import DatasetV2, Subset
-from polaris.experimental._split_v2 import SplitSpecificationV2Mixin
 from polaris.utils.errors import InvalidBenchmarkError
 from polaris.utils.types import ColumnName
 
@@ -14,7 +14,7 @@ class BenchmarkV2Specification(SplitSpecificationV2Mixin, BenchmarkSpecification
     _version: ClassVar[Literal[2]] = 2
 
     dataset: DatasetV2 = Field(exclude=True)
-    n_classes: dict[ColumnName, int]
+    n_classes: dict[ColumnName, int] = Field(default_factory=dict)
 
     @field_validator("dataset", mode="before")
     @classmethod
